07-OSINT-and-Recon/PersonaGraph/requirements.txt
requests==2.31.0
pyyaml==6.0.1
networkx==3.2.1
plotly==5.21.0
beautifulsoup4==4.12.3
numpy==1.26.4
